[FE]: Update Sign Up form
