+++ frontend/package.json
@@ -15,6 +15,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.0",
+    "react-icons": "^5.3.0",
     "react-lazy-load-image-component": "^1.6.2",
     "react-router-dom": "^6.26.2",
     "react-scripts": "5.0.1",
